refactor(audit): extract audit pipeline
